beat/requirements.txt
aio-pika==8.3.0
aioboto3==10.2.0
aiobotocore==2.4.1
aiohttp==3.8.3
aioitertools==0.11.0
aiormq==6.6.4
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.8.1
amqp==5.1.1
anyio==3.6.2
apprise==1.2.0
asgi-lifespan==2.0.0
asks==3.0.0
async-generator==1.10
async-timeout==4.0.2
asyncio-redis-rate-limit==0.2.0
asyncpg==0.27.0
attrs==22.2.0
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickhouse_connect==0.6.4
cloudpickle==2.2.0
colorama==0.4.6
commonmark==0.9.1
coolname==2.0.0
croniter==1.3.7
cryptography==38.0.3
dask==2022.12.0
distributed==2022.12.0
docker==6.0.1
enum34==1.1.10
fastapi==0.87.0
frozenlist==1.3.3
fsspec==2022.11.0
google-auth==2.14.1
greenlet==2.0.1
griffe==0.24.0
h11==0.12.0
h2==4.1.0
HeapDict==1.0.1
hpack==4.0.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==5.0.0
iso8601==1.1.0
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
kombu==5.2.4
kubernetes==25.3.0
locket==1.0.0
loguru==0.6.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
openai==0.27.8
orjson==3.8.1
packaging==21.3
pamqp==3.2.1
pandas==2.0.3
partd==1.3.0
pathspec==0.10.2
pendulum==2.1.2
prefect==2.6.8
prefect-aws==0.1.8
prefect-dask==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.3
redis==4.5.4
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
ruff==0.0.193
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
starlette==0.21.0
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
tortoise-orm==0.19.2
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
websocket-client==1.4.2
websockets==10.4
wrapt==1.14.1
yake==0.4.8
yarl==1.8.2
zict==2.2.0
zipp==3.10.0
